Make views a package
