+++ arches_lingo/views/__init__.py
@@ -1,0 +1,2 @@
+from .root import *
+from .trees import *
