+++ arches_lingo/views/trees.py
@@ -3,10 +3,8 @@ from collections import defaultdict
 from django.contrib.postgres.expressions import ArraySubquery
 from django.db.models import CharField, F, OuterRef, Subquery, Value
 from django.db.models.expressions import CombinedExpression, Func
-from django.shortcuts import render
 from django.utils.translation import gettext_lazy as _
 from django.utils.decorators import method_decorator
-from django.views.decorators.csrf import ensure_csrf_cookie
 from django.views.generic import View
 
 from arches.app.models.models import (
@@ -17,7 +15,6 @@ from arches.app.models.models import (
 )
 from arches.app.utils.decorators import group_required
 from arches.app.utils.response import JSONResponse
-from arches.app.views.base import BaseManagerView
 
 from arches_lingo.const import (
     SCHEMES_GRAPH_ID,
@@ -204,11 +201,3 @@ class ConceptTreeView(View):
         }
         # Todo: filter by nodegroup permissions
         return JSONResponse(data)
-
-
-class LingoRootView(BaseManagerView):
-    @method_decorator(ensure_csrf_cookie)
-    def get(self, request, graphid=None, resourceid=None):
-        context = self.get_context_data(main_script="views/root")
-        context["page_title"] = _("Lingo")
-        return render(request, "arches_lingo/root.htm", context)
